+++ Distroir.CustomSDKLauncher.Core/AppLauncher/AppInfo.cs
@@ -26,6 +26,7 @@ using System.IO;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Drawing.Imaging;
+using Distroir.CustomSDKLauncher.Core.Utilities;
 
 namespace Distroir.CustomSDKLauncher.Core.AppLauncher
 {
