+++ Distroir.CustomSDKLauncher.Core/Backups/BackupEntry.cs
@@ -15,7 +15,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-namespace Distroir.CustomSDKLauncher.Core
+namespace Distroir.CustomSDKLauncher.Core.Backups
 {
     public class BackupEntry
     {
