+++ Distroir.CustomSDKLauncher.Core/Distroir.CustomSDKLauncher.Core.csproj
@@ -61,14 +61,14 @@
     <Compile Include="AppLauncher\Templates\BasicAppTemplate.cs" />
     <Compile Include="AppLauncher\Templates\CustomAppTemplate.cs" />
     <Compile Include="AppLauncher\Templates\JavaAppTemplate.cs" />
-    <Compile Include="BackupEntry.cs" />
-    <Compile Include="BackupManager.cs" />
+    <Compile Include="Backups\BackupEntry.cs" />
+    <Compile Include="Backups\BackupManager.cs" />
     <Compile Include="CommunityContent\ContentData.cs" />
     <Compile Include="CommunityContent\ContentGroup.cs" />
     <Compile Include="CommunityContent\ContentInfo.cs" />
     <Compile Include="CommunityContent\ContentManager.cs" />
     <Compile Include="CommunityContent\ImageData.cs" />
-    <Compile Include="Compressor.cs" />
+    <Compile Include="Utilities\Compressor.cs" />
     <Compile Include="Config.cs" />
     <Compile Include="Feedback\DateSerializer.cs" />
     <Compile Include="Feedback\FeedbackFetcher.cs" />
@@ -76,11 +76,11 @@
     <Compile Include="LanguageManager.cs" />
     <Compile Include="LanguageResourcesList.cs" />
     <Compile Include="Launcher.cs" />
-    <Compile Include="PlatformChecker.cs" />
+    <Compile Include="Utilities\PlatformChecker.cs" />
     <Compile Include="Profile.cs" />
     <Compile Include="ProfileManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="PathFormatter.cs" />
+    <Compile Include="Utilities\PathFormatter.cs" />
     <Compile Include="Template.cs" />
     <Compile Include="TemplateManager.cs" />
     <Compile Include="Tutorial.cs" />
