+++ Distroir.CustomSDKLauncher.Core/Utilities/Compressor.cs
@@ -18,7 +18,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 using System.IO;
 using System.IO.Compression;
 
-namespace Distroir.CustomSDKLauncher.Core
+namespace Distroir.CustomSDKLauncher.Core.Utilities
 {
     public class Compressor
     {
