+++ Distroir.CustomSDKLauncher.Core/Utilities/PlatformChecker.cs
