+++ Distroir.CustomSDKLauncher.UI/Dialogs/SettingsDialog.cs
@@ -18,6 +18,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 using Distroir.Configuration;
 using Distroir.CustomSDKLauncher.Core;
 using Distroir.CustomSDKLauncher.Core.AppLauncher;
+using Distroir.CustomSDKLauncher.Core.Backups;
 using System;
 using System.Collections.Generic;
 using System.Reflection;
